Battle tested, but needs a more formal unit test suite.
This core might still be buggy.  I've noticed sometimes that high bits
are mysteriously set when switching from write to read or vice versa.
Only for one cycle (see
https://github.com/sam-falvo/kestrel/commit/95602c9f290cc5b81db281404c817c6d068dc449 ).
Still not sure what's causing it.  It seems to happen only
intermittently, so I'm wondering if it's really something I can control
in the Verilog somehow.